pyproject.toml
[project]
name = "pathsala-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-role-permissions>=3.2.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
]
